% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_data.R
\name{getCellData}
\alias{getCellData}
\title{Cell Data}
\usage{
getCellData(count_file, separator = ",")
}
\arguments{
\item{count_file}{Path to the csv file that contains the count files}

\item{separator}{The separtor used in the csv file}
}
\value{
A data frame with all the cell counts
}
\description{
Creates a data frame of all the cell data required
}
