% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_functions.R
\name{getCountIndex}
\alias{getCountIndex}
\title{get the start index of a count file, tested on featureCounts only}
\usage{
getCountIndex(cell_data)
}
\arguments{
\item{cell_data}{}
}
\value{
The index of start column.
}
\description{
get the start index of a count file, tested on featureCounts only
}
