% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_obj.R
\name{getMetaFile}
\alias{getMetaFile}
\title{Data Obj}
\usage{
getMetaFile(count_file, separator = ",")
}
\arguments{
\item{count_file}{Path to the csv file that contains the meta data}

\item{separator}{The separtor used in the csv file}
}
\value{
A meta data as Data frame object to be included in the main function
}
\description{
Data Obj
}
