% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_functions.R
\name{prepareBioFardObject}
\alias{prepareBioFardObject}
\title{Create BioFard Object, filter count data and checks meta data}
\usage{
prepareBioFardObject(cell_object, meta_object)
}
\arguments{
\item{cell_object}{Cell Object File}

\item{meta_object}{Meta Object File}
}
\value{
a BioFard Data Object
}
\description{
Create BioFard Object, filter count data and checks meta data
}
